[SRP] Responsabilidades em pacotes e métodos
