+++ cotuba/src/main/java/cotuba/application/Cotuba.java
@@ -1,4 +1,10 @@
-package cotuba;
+package cotuba.application;
+
+import cotuba.domain.Capitulo;
+import cotuba.domain.Ebook;
+import cotuba.epub.GeradorEPUB;
+import cotuba.md.RenderizadorMDParaHTML;
+import cotuba.pdf.GeradorPDF;
 
 import java.nio.file.Path;
 import java.util.List;
