+++ cotuba/src/main/java/cotuba/cli/Main.java
@@ -1,0 +1,39 @@
+package cotuba.cli;
+
+import cotuba.application.Cotuba;
+
+import java.nio.file.Path;
+
+public class Main {
+
+  public static void main(String[] args) {
+
+    Path diretorioDosMD;
+    String formato;
+    Path arquivoDeSaida;
+    boolean modoVerboso = false;
+
+    try {
+
+      LeitorOpcoesCLI opcoesCLI = new LeitorOpcoesCLI(args);
+
+      diretorioDosMD = opcoesCLI.getDiretorioDosMD();
+      formato = opcoesCLI.getFormato();
+      arquivoDeSaida = opcoesCLI.getArquivoDeSaida();
+      modoVerboso = opcoesCLI.isModoVerboso();
+
+      Cotuba cotuba = new Cotuba();
+      cotuba.executa(formato, diretorioDosMD, arquivoDeSaida);
+
+      System.out.println("Arquivo gerado com sucesso: " + arquivoDeSaida);
+
+    } catch (Exception ex) {
+      System.err.println(ex.getMessage());
+      if (modoVerboso) {
+        ex.printStackTrace();
+      }
+      System.exit(1);
+    }
+  }
+
+}
